requirements.txt
fastapi
uvicorn[standard]
google-cloud-spanner
httpx
google-cloud-logging